feat(jsonnet): directly compute imports
Because we are now manually working with the AST, we do not need to use the
TraceImporter anymore, as we have direct access to the importing process.
